TFT_eSPI library required changes
The TFT_eSPI library requires it be configured for each particular build. This is somewhat unusal but that is the way it wa designed. The User_Setup_Select.h file has been modified to include the User_Setups/Setup_Pacman_Clock.h file.
The Setup_Pacman_Clock.h file defines how the TFT is connected and its features (Touch).